cleanup: 64-bit indexing, checked zero-fill, and comment-accuracy items
Mechanical audit leftovers in the non-sampling kernels:

- The scalar-fallback elementwise kernels (unary/binary f1), the axpy
  kernel and both RoPE kernels took 32-bit item indices while their
  vectorized siblings were 64-bit clean - any n or seq*pairs beyond
  INT_MAX silently truncated. All index math is promoted to
  long long (row/position stay int: they are bounded by the vocab-
  scale shapes the kernels serve).
- The GEMV/PC zero-fill paths check their cudaMemsetAsync like every
  other call (a silently failed one would leave torch.empty garbage,
  the exact v0.4 bug class the surrounding comment describes).
- The prefill paged indirection computes table_width * page in
  64-bit and rejects absurd pairs instead of overflowing into
  undefined behavior before the shape check.
- The GQA head mapping uses the group spelling (h / group) the CPU
  references and prefill path use, instead of the equivalent
  cross-multiplied form.
- splitmix_uniform_device / splitmix_uniform_host were hand-synced
  verbatim twins; they are now one __host__ __device__ function.
- The quantize/qadd concurrency contract (process-global scratch,
  distinct-stream calls race it) is documented in cuda_launch.hpp.

+2 tests: the prefill shape matrix gains the LPR=16 band and the
dim <= 512 acceptance boundary (525 total). Performance is neutral by
construction - every change is on an identity path or a K == 0 /
oversize-rejection path; a 3-round probe of topp/rope/softmax
matches the 1.5.0 baselines.

+++ src/fusedtok.cu
@@ -11,8 +11,9 @@ namespace {
 
 // Naive kernel: one thread computes one output element.
 // No memory coalescing tricks, no vectorized loads - correctness first.
-__global__ void axpy_kernel(const float* x, float* y, float a, float b, int n) {
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+__global__ void axpy_kernel(const float* x, float* y, float a, float b,
+                            long long n) {
+    long long i = (long long)blockIdx.x * blockDim.x + threadIdx.x;
     if (i < n) y[i] = a * x[i] + b;
 }
 
@@ -26,7 +27,7 @@ std::vector<float> axpy_cpu(const std::vector<float>& x, float a, float b) {
 
 void axpy_launch(const float* x, float* y, long long n, float a, float b, std::uintptr_t stream) {
     if (n <= 0) return;
-    axpy_kernel<<<(unsigned)grid_for(n), kBlock, 0, (cudaStream_t)stream>>>(x, y, a, b, (int)n);
+    axpy_kernel<<<(unsigned)grid_for(n), kBlock, 0, (cudaStream_t)stream>>>(x, y, a, b, n);
     check_launch("axpy kernel launch");
 }
